auth_service/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
pydantic[email]==2.7.0
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.0
bcrypt==4.1.2
python-dotenv==1.0.0
